requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
requests
beautifulsoup4
langchain-core
langchain-community
langchain-google-genai
mysql-connector-python
psycopg2-binary
alembic
python-multipart
